Moved PreprocessedAggregationFactory to separate file
