+++ aggregators/__init__.py
@@ -1,3 +1,4 @@
 from aggregators.numpy_aggregators import mean, median, trimmed_mean
-from aggregators.numpy_aggregator_factory import NumpyAggregationFactory, PreprocessedAggregationFactory
+from aggregators.numpy_aggregation_factory import NumpyAggregationFactory
+from aggregators.preprocessed_aggregation_factory import PreprocessedAggregationFactory
 from aggregators.robust_weiszfeld_factory import RobustWeiszfeldFactory
